Um Dao para cada tabela da db

+++ app/src/main/java/com/example/gestaodacozinha/data/ProdutoQuantidade.kt
@@ -7,6 +7,6 @@ import androidx.room.Entity
     primaryKeys = ["produtoId", "quantidade"]
 )
 data class ProdutoQuantidade(
-    val produtoId: Float,
+    val produtoId: Long,
     val quantidade: Float,
 )
